Refactor User in Ui
