+++ presentation/src/main/java/com/pocs/presentation/mock/Mocks.kt
@@ -4,7 +4,7 @@ import androidx.paging.PagingData
 import com.pocs.domain.model.user.UserDetail
 import com.pocs.domain.model.user.UserType
 import com.pocs.presentation.model.post.item.PostItemUiState
-import com.pocs.presentation.model.user.UserItemUiState
+import com.pocs.presentation.model.user.item.UserItemUiState
 
 val mockNoticeItemsPagingData = PagingData.from(
     listOf(
