+++ presentation/src/main/java/com/pocs/presentation/model/user/UserDetailUiState.kt
@@ -1,9 +1,9 @@
 package com.pocs.presentation.model.user
 
-import com.pocs.domain.model.user.UserDetail
+import com.pocs.presentation.model.user.item.UserDetailItemUiState
 
 sealed class UserDetailUiState {
-    class Success(val userDetail: UserDetail) : UserDetailUiState()
+    class Success(val userDetail: UserDetailItemUiState) : UserDetailUiState()
     class Failure(val e: Throwable, val onRetryClick: () -> Unit) : UserDetailUiState()
     object Loading : UserDetailUiState()
 }
