+++ presentation/src/main/java/com/pocs/presentation/model/user/item/UserItemUiState.kt
@@ -1,4 +1,4 @@
-package com.pocs.presentation.model.user
+package com.pocs.presentation.model.user.item
 
 data class UserItemUiState(
     val id: Int,
